Refactor models into dedicated module
