--- app1.py
@@ -1,4 +1,0 @@
-# models.py
-from app import db
-from flask_login import UserMixin
-from werkzeug.security import generate_password_ha
